+++ libs/email-service/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/types';
-export * from './lib/abstract';
-export * from './lib/local-file-emails';
+export * from './lib/abstract-email-service';
+export * from './lib/local-file-email-service';
+export * from './lib/noop-email-service';
