+++ libs/email-service/src/lib/abstract-email-service.ts
